fix: move build test to integration tier and fix --help exit code
Move TestMakeBuildMCP to build_test.go with //go:build integration tag
since it shells out to make and runs a compiled binary. Fix incorrect
assumption that --help exits with code 2 (flag.ErrHelp) — Go's flag
package actually exits with code 0 for --help.
